backend/requirements.txt
Django==4.1.7
django-cors-headers==3.14.0
django-dotenv==1.4.2
django-environ==0.10.0
django-twilio==0.14.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
psycopg2==2.9.5